relay/requirements.txt
# FlowClaw Relay Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.0
cryptography==42.0.0
pynacl==1.5.0
openai==1.40.0
